name genarator function moved to helper folder.
